AMP-1902
git-svn-id: file:///svn/trunk@3117 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/action/ViewNewAdvancedReport.java
@@ -69,7 +69,7 @@ public class ViewNewAdvancedReport extends Action {
 		
 		if(ampAppSettings!=null)
 			request.setAttribute("recordsPerPage", ampAppSettings.getDefaultRecordsPerPage());
-		else request.setAttribute("recordsPerPage", new Integer(10));
+		else request.setAttribute("recordsPerPage", new Integer(25));
 		
 		
 		//check currency code:
